Added  button and function for add new transaction
