+++ personal_expense_app/lib/main.dart
@@ -1,15 +1,48 @@
 import 'package:flutter/material.dart';
+import 'package:personal_expense_app/widgets/new_transaction.dart';
+import 'package:personal_expense_app/widgets/transcation_list.dart';
 import './models/transcation.dart';
-import './widgets/user_transactions.dart';
 
 void main() => runApp(MyApp());
 
-final titleController = TextEditingController();
-final amountController = TextEditingController();
 String titleInput;
 String amountInput;
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
+  @override
+  _MyAppState createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  final titleController = TextEditingController();
+  final amountController = TextEditingController();
+
+  final List<Transaction> _userTransaction = [
+    Transaction(
+        id: 't1', title: 'New Clothes', amount: 599, date: DateTime.now()),
+    Transaction(id: 't2', title: 'Pendrive', amount: 1899, date: DateTime.now())
+  ];
+
+  void _addNewTransaction(String txTitle, double txAmount) {
+    final newTx = Transaction(
+        title: txTitle,
+        amount: txAmount,
+        date: DateTime.now(),
+        id: DateTime.now().toString());
+
+    setState(() {
+      _userTransaction.add(newTx);
+    });
+  }
+
+  void startAddNewTransction(BuildContext ctxt) {
+    showModalBottomSheet(
+        context: ctxt,
+        builder: (bCtxt) {
+          return NewTransaction(_addNewTransaction);
+        });
+  }
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -24,33 +57,37 @@ class MyHomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: Text('Personal Expense'),
-        ),
-        body: SingleChildScrollView(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: <Widget>[
-              Container(
-                width: 150,
-                height: 100,
-                child: Card(
-                  shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(20)),
-                  shadowColor: Colors.black,
-                  color: Colors.deepPurpleAccent,
-                  child: Text(
-                    '  Some Dummy Text for Chart Widget',
-                    style: TextStyle(color: Colors.white, fontSize: 20),
-                  ), // width of the card is depnded on its child. Unless it itself is a child like in this case we have made CARD a child of the contianer and therefore it listens to the parent. Bydefault it takes as much width as it is required for the child/parent.
-                  elevation:
-                      20, // TL;DR card takes childs width unless there is a width defined by the parent then it takes the parents width
-                ),
+      appBar: AppBar(
+        title: Text('Personal Expense'),
+        actions: <Widget>[IconButton(icon: Icon(Icons.add), onPressed: () {})],
+      ),
+      body: SingleChildScrollView(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.start,
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: <Widget>[
+            Container(
+              width: 150,
+              height: 100,
+              child: Card(
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(20)),
+                shadowColor: Colors.black,
+                color: Colors.deepPurpleAccent,
+                child: Text(
+                  '  Some Dummy Text for Chart Widget',
+                  style: TextStyle(color: Colors.white, fontSize: 20),
+                ), // width of the card is depnded on its child. Unless it itself is a child like in this case we have made CARD a child of the contianer and therefore it listens to the parent. Bydefault it takes as much width as it is required for the child/parent.
+                elevation:
+                    20, // TL;DR card takes childs width unless there is a width defined by the parent then it takes the parents width
               ),
-              UserTransactions(),
-            ],
-          ),
-        ));
+            ),
+            TransactionList(transactions),
+          ],
+        ),
+      ),
+      floatingActionButton:
+          FloatingActionButton(child: Icon(Icons.add), onPressed: () {}),
+    );
   }
 }
