+++ personal_expense_app/lib/widgets/new_transaction.dart
@@ -4,6 +4,18 @@ class NewTransaction extends StatelessWidget {
   final Function
       addTx; // we created this funtion so that we could pass this on press of the add transcation button
   NewTransaction(this.addTx);
+
+  void onSubmit() {
+    final enteredTitle = titleController.text;
+    final enteredAmount = double.parse(amountController.text);
+
+    if (enteredTitle.isEmpty || enteredTitle.isEmpty) {
+      return;
+    }
+
+    addTx(enteredTitle, enteredAmount);
+  }
+
   final titleController = TextEditingController();
   final amountController = TextEditingController();
   @override
@@ -19,27 +31,16 @@ class NewTransaction extends StatelessWidget {
                 labelText: 'Title',
                 hintText: 'Enter the Description of your purchase'),
             controller: titleController,
-            // onChanged: (val) {
-            //   titleInput = val;
-            //   print(titleInput);
-            // },
           ),
           TextField(
             decoration: InputDecoration(
                 labelText: 'Amount', hintText: 'Enter the amount'),
             controller: amountController,
-            // onChanged: (value) {
-            //   amountInput = value;
-            // },
             keyboardType: TextInputType.number,
+            onSubmitted: (_) => onSubmit(),
           ),
           FlatButton(
-              onPressed: () {
-                addTx(
-                    titleController.text, double.parse(amountController.text));
-                print(titleController.text);
-                print(amountController.text);
-              },
+              onPressed: onSubmit,
               textColor: Colors.deepPurpleAccent,
               child: Text('Add Transaction'))
         ],
